refactor(oparl): split derivate types from Zod schemas (Astro-only)
Keep Zod off the Deno side. oparl-derivatives.ts is now an import-free
types module — the file the Deno workspace reexports — so `deno check` no
longer resolves Zod and `npm:zod` drops out of deno.json (and the Astro
package.json entry added earlier is reverted; the schema now uses astro/zod).

The real Zod schemas move to a new Astro-only oparl-derivatives.schema.ts
(not reexported to Deno), keeping the spec §5.1 fail-fast validation intact
for content.config to consume in PR A/B. A compile-time drift guard
(`AssertEqual<z.infer<typeof schema>, Type>`) fails the Astro build if a
schema and its canonical type diverge — verified it bites on a mismatch.

Records the decision and its primary-source basis in
docs/research/2026-07-12-oparl-derivatives-zod-alternatives.md and relaxes
spec §3 (one file with both -> types file + Astro-only schema file).
