+++ astro/package.json
@@ -35,8 +35,7 @@
     "tailwindcss": "^4.1.4",
     "typescript": "^5.8.3",
     "typesense": "^3.0.6",
-    "vitest": "^4.0.4",
-    "zod": "^4.3.6"
+    "vitest": "^4.0.4"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.16",
