+++ deno.lock
@@ -38,8 +38,7 @@
     "npm:openai@*": "5.6.0_zod@4.4.3",
     "npm:openai@^5.6.0": "5.6.0_zod@4.4.3",
     "npm:tesseract.js@*": "6.0.1",
-    "npm:tesseract.js@^6.0.1": "6.0.1",
-    "npm:zod@^4.3.6": "4.4.3"
+    "npm:tesseract.js@^6.0.1": "6.0.1"
   },
   "jsr": {
     "@denosaurs/plug@1.0.5": {
@@ -615,8 +614,7 @@
       "jsr:@std/path@^1.0.8",
       "jsr:@std/testing@^1.0.5",
       "npm:@aws-sdk/client-s3@3",
-      "npm:openai@^5.6.0",
-      "npm:zod@^4.3.6"
+      "npm:openai@^5.6.0"
     ]
   }
 }
